remove usage of global variable
