+++ autoemulate/emulators/neural_networks/base.py
@@ -3,17 +3,6 @@ from torch import nn
 
 from autoemulate.utils import set_random_seed
 
-_MODULES = dict()
-
-
-def register(name):
-    def add_to_dict(fn):
-        global _MODULES
-        _MODULES[name] = fn
-        return fn
-
-    return add_to_dict
-
 
 class TorchModule(nn.Module):
     """
@@ -40,16 +29,3 @@ class TorchModule(nn.Module):
 
     def forward(self, X: torch.Tensor):
         raise NotImplementedError("forward method not implemented.")
-
-
-def get_module(module: str | TorchModule, module_args) -> TorchModule:
-    """
-    Return the module instance for NeuralNetRegressor. If `module` is a string,
-    then initialize a TorchModule with the same registered name. If `module` is
-    already a TorchModule, then return it as is.
-    """
-    if isinstance(module, TorchModule):
-        return module
-    if module not in _MODULES:
-        raise NotImplementedError(f"Module {module} not implemented.")
-    return _MODULES[module](**module_args)
